Store block type to id mapping to fix compatibility with 1.13. (#51)
This solves issues with legacy materials not being available at runtime (which breaks the plugin in 1.13) and them potentially getting removed in the future.
On first startup it will create a new types table and tries to populate it with the legacy type IDs where possible. For the new/non-matchable types it will generate arbitrary ID numbers.